+++ crates/masm-analysis/src/unconstrained_advice/mod.rs
@@ -3,6 +3,7 @@
 mod address;
 mod call_transfer;
 mod domain;
+mod env;
 mod grouping;
 mod inter;
 mod merkle;
